requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.11
black==25.1.0
certifi==2025.4.26
click==8.2.1
dill==0.4.0
distro==1.9.0
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isort==6.0.1
jiter==0.10.0
mccabe==0.7.0
mypy_extensions==1.1.0
openai==1.84.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pydantic==2.11.5
pydantic_core==2.33.2
pylint==3.3.7
python-dotenv==1.1.0
sniffio==1.3.1
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
